Change Windows multiprocessing logic so that it's more intuitive: you (#100)
set the -j1 flag if you DO NOT want it to be parallel. Otherwise on
Windows it will default to parallel builds (just like on Linux and OSX)

If there is anything in the CL environment variable, append to it rather
than overwrite. Ensure that we aren't overwriting an existing /MP usage
in the CL environment variable.